requirements.txt
neo4j==5.28.1
python-dotenv==1.0.1
tiktoken==0.9.0
pypdf==5.4.0
tiktoken==0.9.0
openai==1.70.0
google-cloud-bigquery==3.14.1
google-cloud-bigquery-storage==2.26.0
google-generativeai==0.8.3
groq==0.13.0
tenacity==9.0.0
PyYAML==6.0.2